requirements.txt
web3>=7.0.0
py-solc-x>=2.0.0
python-dotenv>=1.0.1
aiohttp>=3.10.0
